requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages

astronomer-cosmos[dbt-bigquery]==1.3.2
protobuf==3.20.0